更新 layui 修复列表排序，分页，搜索问题

+++ models/Common.go
@@ -5,8 +5,8 @@ import "BeeCustom/enums"
 // JsonResult 用于返回ajax请求的基类
 type JsonResult struct {
 	Status enums.JsonResultCode `json:"status"`
-	Msg  string               `json:"msg"`
-	Obj  interface{}          `json:"obj"`
+	Msg    string               `json:"msg"`
+	Obj    interface{}          `json:"obj"`
 }
 
 // BaseQueryParam 用于查询的类
@@ -14,5 +14,5 @@ type BaseQueryParam struct {
 	Sort   string `json:"sort"`
 	Order  string `json:"order"`
 	Offset int64  `json:"offset"`
-	Limit  int    `json:"limit"`
+	Limit  int64  `json:"limit"`
 }
